chore: begin advisory lock rewrite.
