+++ Cargo.toml
@@ -80,6 +80,7 @@ trybuild = { git = "https://github.com/l-monninger/trybuild", rev = "21f376043b1
 git2 = "0.18"
 glob = "0.3"
 pathdiff = "0.2"
+file-lock = { version = "2.1.11" }
  
 # internal
 cite-core = { path = "cite/core" }
