+++ cite/git/src/repository/repository.rs
@@ -1,0 +1,1 @@
+pub struct Repository {}
